Commented the code for traceability. As of 10/8/14 the following requirements have yet to be completed for iteration 1:
1.4.4: The current turn shall be indicated by highlighting that turn on
the left side of the screen.

1.4.5: The current highest achieved score shall be displayed. This score
shall initially be set to 5000 points.

1.6.0: At the conclusion of the game, the user shall be greeted with
three options: “Play again?”, “Main Menu”, and “Quit”.

2.1.0: (Turn highlighting only) When 1 player mode is selected, the 1
player mode graphic user interface is displayed with blank dice, the
bank button disabled, and turn number one highlighted. The user will
have ten turns to try and get as many points as possible.

2.1.2: (High scores only) The game ends at the conclusion of the tenth
turn, and the player’s score is compared to the current high score.

2.1.3: If the player’s score is greater than the current high score, a
congratulatory message is displayed, and the player’s score replaces the
previous high score.

2.1.4: The player is given three options: “Play again?”, “Main Menu”,
and “Quit”.

2.1.5: If “Play again?” is selected, the game starts over in 1 player
mode.

2.1.6: If “Main Menu” is selected, the user is taken to the mode
selection screen.

2.1.7: If “Quit” is selected, the application immediately closes.

In total, it looks like we're missing 10 requirements out of around 45
or so for iteration 1. That puts us at ~ 78%. Realistically, I think
we're closer than that. Idk if we want to move some of these to
iteration 2 or what. Maybe we could talk Saturday.